requirements.txt
coverage == 4.5.4
python-decouple == 3.4
Django == 3.1.7
django-cors-headers == 3.7.0
django-dotenv == 1.4.2
django-filter == 2.4.0
django-fontawesome-5 == 1.0.18
django-rest-knox == 4.1.0
djangorestframework == 3.12.2
mardown == 3.3.4
mod-wsgi == 4.6.8
postgres == 3.0.0
pytz == 2019.1
sqlparse == 0.3.0
wheel == 0.36.2
